<FCV2> Changed Flight Controller to Attitude Rate Mode + Used Transmitter Channel 5 to Tune PID
Changes:
1. Implemented attitude rate mode in order to tune PID.
2. Changed scheme to enter PID tuning mode. Now if channel 5 of controller is maximum, FC will enter tuning PID mode.

Test & Result:
No Error

+++ FlightController/FlightControllerV2/Src/drivers/LED/led.c
@@ -47,7 +47,7 @@ bool LED_Toggle(LEDType led)
 bool LED_Blink(LEDType led, uint8_t cnt)
 {
     // we blink every 200ms
-    for (int i = 0; i < cnt; ++i) {
+    for (uint8_t i = 0; i < cnt; ++i) {
         LED_Toggle(led);
         HAL_Delay(200);
     }
